PlatformConfig: Build init_boot image
Builds a new init_boot.img with the ramdisk that was previously in
boot.img. For devices launching with Android 13, the generic ramdisk
is removed from the boot image and placed in a separate init_boot
image [1]. This change leaves the boot image with only the kernel.

Please note that the fastboot supplied with the bootloader does not
know how to flash the init_boot image and this must be done from the
fastbootd!

[1] https://source.android.com/docs/core/architecture/partitions/generic-boot

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>